Breaks apart WebCheckoutViewController into Checkout.* namespace and folder on disk
