+++ Assets/Plugins/iOS/Shopify/Checkout/Checkout.State.swift
@@ -1,0 +1,58 @@
+//
+//  Checkout.State.swift
+//  UnityBuySDK
+//
+//  Created by Shopify.
+//  Copyright © 2017 Shopify Inc. All rights reserved.
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import UIKit
+
+extension Checkout {
+    
+enum State {
+    case checkingOut
+    case finished
+    
+    // Colors from https://polaris.shopify.com/visuals/colors#color-usage
+    var highlightColor: UIColor {
+        switch self {
+        case .checkingOut: return UIColor(hex: "FEAF9A")!
+        case .finished: return UIColor(hex: "BBE5B3")!
+        }
+    }
+    
+    var normalColor: UIColor {
+        switch self {
+        case .checkingOut: return UIColor(hex: "ED6347")!
+        case .finished: return UIColor(hex: "50B83C")!
+        }
+    }
+    
+    var title: String {
+        switch self {
+        case .checkingOut: return "Cancel"
+        case .finished: return "Finish"
+        }
+    }
+}
+    
+}
